Added plate polygons to client library
