+++ corelle-client/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./feature";
 export * from "./provider";
+export * from "./plate-polygons";
